Draft for running integration tests
