+++ config/witness-demo/wan.json
@@ -1,0 +1,8 @@
+{
+    "dt": "2022-01-20T12:57:59.823350+00:00",
+    "wan": {
+        "dt": "2022-01-20T12:57:59.823350+00:00",
+        "curls": ["tcp://witness-demo:5632/", "http://witness-demo:5642/"]
+    },
+    "iurls": []
+}
