+++ config/witness-demo/wes.json
@@ -1,0 +1,8 @@
+{
+    "wes": {
+        "dt": "2022-01-20T12:57:59.823350+00:00",
+        "curls": ["tcp://witness-demo:5634/", "http://witness-demo:5644/"]
+    },
+    "dt": "2022-01-20T12:57:59.823350+00:00",
+    "iurls": []
+}
