+++ config/witness-demo/wil.json
@@ -1,0 +1,8 @@
+{
+    "wil": {
+        "dt": "2022-01-20T12:57:59.823350+00:00",
+        "curls": ["tcp://witness-demo:5633/", "http://witness-demo:5643/"]
+    },
+    "dt": "2022-01-20T12:57:59.823350+00:00",
+    "iurls": []
+}
