+++ config/witness-demo/wit.json
@@ -1,0 +1,8 @@
+{
+    "wit": {
+        "dt": "2022-01-20T12:57:59.823350+00:00",
+        "curls": ["tcp://witness-demo:5635/", "http://witness-demo:5645/"]
+    },
+    "dt": "2022-01-20T12:57:59.823350+00:00",
+    "iurls": []
+}
