+++ config/witness-demo/wub.json
@@ -1,0 +1,8 @@
+{
+    "wub": {
+        "dt": "2022-01-20T12:57:59.823350+00:00",
+        "curls": ["tcp://witness-demo:5636/", "http://witness-demo:5646/"]
+    },
+    "dt": "2022-01-20T12:57:59.823350+00:00",
+    "iurls": []
+}
