+++ config/witness-demo/wyz.json
@@ -1,0 +1,8 @@
+{
+    "wyz": {
+        "dt": "2022-01-20T12:57:59.823350+00:00",
+        "curls": ["tcp://witness-demo:5637/", "http://witness-demo:5647/"]
+    },
+    "dt": "2022-01-20T12:57:59.823350+00:00",
+    "iurls": []
+}
