+++ docker-compose.yaml
@@ -1,0 +1,76 @@
+services:
+    vlei-server:
+        image: gleif/vlei
+        command:
+            - vLEI-server
+            - -s
+            - ./schema/acdc
+            - -c
+            - ./samples/acdc/
+            - -o
+            - ./samples/oobis/
+        healthcheck:
+            test:
+                - CMD
+                - curl
+                - -f
+                - http://localhost:7723/oobi/EBfdlu8R27Fbx-ehrqwImnK-8Cm79sqbAQ4MmvEAYqao
+            interval: 2s
+            timeout: 3s
+            retries: 5
+            start_period: 2s
+        ports:
+            - 7723:7723
+
+    keria:
+        image: weboftrust/keria:latest
+        environment:
+            - KERI_AGENT_CORS=1
+            - KERI_URL=http://keria:3902
+        volumes:
+            - ./config/keria.json:/keria/config/keri/cf/keria.json
+        entrypoint: keria
+        command:
+            - start
+            - --config-dir
+            - /keria/config
+            - --config-file
+            - keria
+            - --name
+            - agent
+        healthcheck:
+            test: ['CMD', 'curl', '-f', 'http://localhost:3902/spec.yaml']
+            interval: 2s
+            timeout: 3s
+            retries: 5
+            start_period: 2s
+        ports:
+            - 3901:3901
+            - 3902:3902
+            - 3903:3903
+
+    witness-demo:
+        image: weboftrust/keri-witness-demo:1.1.0
+        healthcheck:
+            test: ['CMD', 'curl', '-f', 'http://localhost:5642/oobi']
+            interval: 2s
+            timeout: 3s
+            retries: 5
+            start_period: 2s
+        volumes:
+            - ./config/witness-demo:/keripy/scripts/keri/cf/main
+        ports:
+            - 5642:5642
+            - 5643:5643
+            - 5644:5644
+
+    deps:
+        image: alpine
+        command: ['echo', 'Dependencies running']
+        depends_on:
+            vlei-server:
+                condition: service_healthy
+            keria:
+                condition: service_healthy
+            witness-demo:
+                condition: service_healthy
