+++ package.json
@@ -18,9 +18,10 @@
     "start": "npm run build:esm -- --watch",
     "build:cjs": "esbuild src/index.ts --bundle --outfile=dist/signify-ts.cjs --platform=neutral --format=cjs --sourcemap --packages=external",
     "build:esm": "esbuild src/index.ts --bundle --outfile=dist/signify-ts.mjs --platform=neutral --format=esm --sourcemap --packages=external",
-    "build:types": "tsc --declaration --outDir dist",
+    "build:types": "tsc -p tsconfig.build.json",
     "build": "npm run build:cjs && npm run build:esm && npm run build:types",
-    "test": "jest",
+    "test": "jest test/",
+    "test:integration": "jest examples/integration-scripts/",
     "lint": "npx eslint src test examples/integration-scripts examples/scripts",
     "prepare": "npm run build",
     "generate-docs": "node_modules/.bin/jsdoc --configure .jsdoc.json --verbose",
