+++ tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+    "include": ["src", "types"],
+    "extends": "./tsconfig.node.json",
+    "compilerOptions": {
+        "outDir": "dist",
+        "declaration": true,
+        "emitDeclarationOnly": true
+    }
+}
