+++ tsconfig.json
@@ -1,9 +1,8 @@
 {
-  "include": ["src", "types"],
+  "include": ["src", "types", "test", "examples/integration-scripts"],
   "extends": "./tsconfig.node.json",
   "compilerOptions": {
-    "outDir": "dist",
-    "declaration": true,
-    "emitDeclarationOnly": true
+    "noEmit": true,
+    "allowImportingTsExtensions": true
   }
 }
